Add files via upload
The dataset is in a CSV format and every row represents one person stopped by a police officer. I cleaned the data using Python, 
and limited the number of observations to the first 10,000 records with complete data. I dropped the observations that
had any missing information. There are no aggregations needed to this dataset. The front end requires variables with complete 
values.